requirements.txt
--index-url https://download.pytorch.org/whl/cu130
torch==2.9.1+cu130
torchvision==0.24.1+cu130
torchaudio==2.9.1+cu130
numpy>=1.24
scikit-learn>=1.2
lightgbm>=4.0
onnx>=1.15
onnxruntime-gpu>=1.15
skl2onnx>=1.15
joblib>=1.3
matplotlib>=3.6
sentence-transformers>=2.2
transformers>=4.35.0
Pillow>=9.5
tqdm>=4.64.0
jupyter_client>=8.0.0
ipykernel>=6.0.0
pytest>=7.2
flask>=2.2
onnxmltools>=1.11
onnxruntime>=1.15 
psutil
sentencepiece